All warnings fixed and made clicks time based

+++ trunk/Konstruct/Client/Dimensions/AreaSpray.cpp
@@ -28,7 +28,7 @@ bool AreaSpray::Activate(PlayerCharacter *pSkillOwner)
 		Weapon* pEquipped = pSkillOwner->GetEquippedWeapon();
 		m_fRange = pEquipped->GetRange() * m_fRangeMod;
 		m_iShotsToFire = MIN_SHOTS + m_iSkillRank / m_iShotsMod;
-		m_iDamage = (pEquipped->GetDamage() * m_fDamageMod) + (iRankMultiple * m_iSkillRank);
+		m_iDamage = int(pEquipped->GetDamage() * m_fDamageMod) + (iRankMultiple * m_iSkillRank);
 		m_eDamageType = pEquipped->GetDamageType();
 		m_fSpeed = pEquipped->GetSpeed() * m_fSpeedMod;
 		m_fRecovery = pEquipped->GetRecovery() * m_fRecoveryMod;
